refactor based on previous change
