+++ src/main/java/frc/team5115/Main.java
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot;
+package frc.team5115;
 
 import edu.wpi.first.wpilibj.RobotBase;
