+++ src/main/java/frc/team5115/commands/SpinAmper.java
@@ -1,8 +1,8 @@
-package frc.robot.commands;
+package frc.team5115.commands;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.shooter.Shooter;
+import frc.team5115.subsystems.shooter.Shooter;
 
 public class SpinAmper extends Command {
     private final Shooter shooter;
