+++ src/main/java/frc/team5115/subsystems/arm/ArmIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.arm;
+package frc.team5115.subsystems.arm;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import org.littletonrobotics.junction.AutoLog;
