+++ src/main/java/frc/team5115/subsystems/shooter/Shooter.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.shooter;
+package frc.team5115.subsystems.shooter;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.controller.SimpleMotorFeedforward;
@@ -6,7 +6,9 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
+import frc.robot.subsystems.shooter.ShooterIOInputsAutoLogged;
+import frc.team5115.Constants;
+
 import org.littletonrobotics.junction.Logger;
 
 public class Shooter extends SubsystemBase {
