+++ src/main/java/frc/team5115/subsystems/shooter/ShooterIOSim.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.shooter;
+package frc.team5115.subsystems.shooter;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.geometry.Rotation2d;
